Format files

+++ src/solutions/easy/two_sum.rs
@@ -5,7 +5,7 @@ pub fn two_sum(nums: Vec<i32>, target: i32) -> Vec<i32> {
     for (i, &num) in nums.iter().enumerate() {
         let delta = target - num;
         if let Some(&index) = delta_hashmap.get(&delta) {
-            return vec![index as i32, i as i32];
+            return vec![index, i as i32];
         }
         delta_hashmap.insert(num, i as i32);
     }
